+++ tests/examples/examples_test.go
@@ -188,35 +188,34 @@ func TestAccHelmLocal(t *testing.T) {
 	integration.ProgramTest(t, &test)
 }
 
-// TODO: uncomment once https://github.com/pulumi/pulumi-kubernetes/issues/1137 is fixed.
-//func TestAccPrometheusOperator(t *testing.T) {
-//	skipIfShort(t)
-//	test := getBaseOptions(t).
-//		With(integration.ProgramTestOptions{
-//			Dir:         path.Join(getCwd(t), "prometheus-operator"),
-//			SkipRefresh: true,
-//			ExtraRuntimeValidation: func(
-//				t *testing.T, stackInfo integration.RuntimeValidationStackInfo,
-//			) {
-//				assert.NotNil(t, stackInfo.Deployment)
-//				assert.Equal(t, 10, len(stackInfo.Deployment.Resources))
-//			},
-//			EditDirs: []integration.EditDir{
-//				{
-//					Dir:      path.Join(getCwd(t), "prometheus-operator", "steps"),
-//					Additive: true,
-//					ExtraRuntimeValidation: func(
-//						t *testing.T, stackInfo integration.RuntimeValidationStackInfo,
-//					) {
-//						assert.NotNil(t, stackInfo.Deployment)
-//						assert.Equal(t, 10, len(stackInfo.Deployment.Resources))
-//					},
-//				},
-//			},
-//		})
-//
-//	integration.ProgramTest(t, &test)
-//}
+func TestAccPrometheusOperator(t *testing.T) {
+	skipIfShort(t)
+	test := getBaseOptions(t).
+		With(integration.ProgramTestOptions{
+			Dir:         path.Join(getCwd(t), "prometheus-operator"),
+			SkipRefresh: true,
+			ExtraRuntimeValidation: func(
+				t *testing.T, stackInfo integration.RuntimeValidationStackInfo,
+			) {
+				assert.NotNil(t, stackInfo.Deployment)
+				assert.Equal(t, 10, len(stackInfo.Deployment.Resources))
+			},
+			EditDirs: []integration.EditDir{
+				{
+					Dir:      path.Join(getCwd(t), "prometheus-operator", "steps"),
+					Additive: true,
+					ExtraRuntimeValidation: func(
+						t *testing.T, stackInfo integration.RuntimeValidationStackInfo,
+					) {
+						assert.NotNil(t, stackInfo.Deployment)
+						assert.Equal(t, 10, len(stackInfo.Deployment.Resources))
+					},
+				},
+			},
+		})
+
+	integration.ProgramTest(t, &test)
+}
 
 func TestAccMariadb(t *testing.T) {
 	skipIfShort(t)